GH-87: fixed comments
