+++ vector/src/test/java/org/apache/arrow/vector/TestStructVector.java
@@ -42,7 +42,7 @@ import org.apache.arrow.vector.types.pojo.ArrowType;
 import org.apache.arrow.vector.types.pojo.ArrowType.Struct;
 import org.apache.arrow.vector.types.pojo.Field;
 import org.apache.arrow.vector.types.pojo.FieldType;
-import org.apache.arrow.vector.types.pojo.TestUuidVector;
+import org.apache.arrow.vector.types.pojo.UuidType;
 import org.apache.arrow.vector.util.TransferPair;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -341,7 +341,7 @@ public class TestStructVector {
 
   @Test
   public void testStructVectorWithExtensionTypes() {
-    TestUuidVector.UuidType uuidType = new TestUuidVector.UuidType();
+    UuidType uuidType = new UuidType();
     Field uuidField = new Field("struct_child", FieldType.nullable(uuidType), null);
     Field structField =
         new Field("struct", FieldType.nullable(new ArrowType.Struct()), List.of(uuidField));
@@ -353,14 +353,14 @@ public class TestStructVector {
 
   @Test
   public void testStructVectorTransferPairWithExtensionType() {
-    TestUuidVector.UuidType uuidType = new TestUuidVector.UuidType();
+    UuidType uuidType = new UuidType();
     Field uuidField = new Field("uuid_child", FieldType.nullable(uuidType), null);
     Field structField =
         new Field("struct", FieldType.nullable(new ArrowType.Struct()), List.of(uuidField));
 
     StructVector s1 = (StructVector) structField.createVector(allocator);
-    TestUuidVector.UuidVector uuidVector =
-        s1.addOrGet("uuid_child", FieldType.nullable(uuidType), TestUuidVector.UuidVector.class);
+    UuidVector uuidVector =
+        s1.addOrGet("uuid_child", FieldType.nullable(uuidType), UuidVector.class);
     s1.setValueCount(1);
     uuidVector.set(0, new UUID(1, 2));
     s1.setIndexDefined(0);
